Removed org-team and org-collaborator relation
